+++ NEWS.md
@@ -1,3 +1,10 @@
+# modeltime.ensemble (development version)
+
+__Stacked Ensembles__
+
+- Use `modeltime_fit_resamples()` to generate resampled predictions
+- Use `ensemble_model_spec()` to apply stacking using a `model_spec`
+
 # modeltime.ensemble 0.1.0
 
 * Initial release of `modeltime.ensemble`.